[aptos-framework] Make Tokens unique by name
While the IDs are very elegant and safe, the issue is that there's no
convenient method within Aptos to get those IDs without scraping all the
user data or events -- but then there might be multiple tokens or
collections within a creators account with the same name. The only way
to really resolve this is to introduce uniqueness from the creators
side.

GitOrigin-RevId: 7288a6f085a038a2cbffc3385f12bf60fba7ba56