[framework] support type_of
which will be helpful to elimiante points of centralization
namely one can create new coin standards and other concepts without
relying on a system account or paying for arbitrary folks to store stuff
on their account to lock in strong uniqueness properties offered by
types.

this is crucial for a general purpose coin.

GitOrigin-RevId: b02301331406150f14994d984a019f575f8a3ed7