[framework] remove validator_set_script
This was a legacy API used by the Diem framework for managing
validators. In the interesting of preparing our framework for V1, it is
time to kill it.

Note, state sync v1, which is deprecated, has tests that use this for
boilerplate transactions. Hence the awkward executor_proxy code.

Also the team has intended to keep around the operator / management
codebase until there are smoke tests covering the new CLI tool. Hence
the motivation for the awkward unimplemented macros.

GitOrigin-RevId: 9d689ef0709c59121877dba9cc579c3560280a82