[framework] support a table that doesn't track length
a table with length that has a lot of adds or removals won't work well
for block-stm, so create a table that doesn't track length for optimal
concurrency.

this becomes the default table, then wrap it with a table that has
length and can be destroyed... it leaves open the question about systems
like token_transfers which wants to use tables... but the flip side is
that concurrency is irrelevant for a single account in the current model.

GitOrigin-RevId: 4f68097e2ba8869e02f5909a40ca268d71069557