[code upgrade] Verify package dependency policy (#3448)
A package is not allowed to refer to another package which has a lower upgrade policy. For example, an immutable package cannot refer to a compatible package. This check wasn't yet implemented and is added in this PR. It generally makes the policy system more sound as the upgrade police guarantees of a given package now extend to it's transitive closure of dependencies.

This PR also suggests to renable policy `arbitrary` with a semantics such that is suited for the 'app' developer (in contrast to the library developer). Such a developer may want to have more freedom on upgrading during initial development. Those developers can build an interdependent system of packages on a _single address_. Thus, a dependency to an `arbitrary` package is now only allowed if that package resides at the same address.

The PR is an example for a first framework-only upgrade. Changes to on-chain data representation are compatible (via `Any`), and no new native logic is added to validator side. On SDK side, the tools for building packages and generating metadata needed to be updated. This can be bundled with other, similar PRs in an upcoming SDK/framework upgrade.

GitOrigin-RevId: f53eddd6fd3722ccec3eb48dc9aa403a3cd7cf7a