[vm] add a native test function for reading events
this is useful for ensuring that emitted events match expectations.

an increasingly large number of developers have become dependent on the
correctness of events. currently the only way to test events is at an
integration layer, which creates challenges for pure smart contract
teams. this change will improve their developer experience
substantially.

GitOrigin-RevId: dfc012af8aa04b191355e3c4a3d9a1c9ce67f6d1