[aptos-vm] enable automatic creation of sponsored accounts
This is the implementation of AIP-52:
https://github.com/aptos-foundation/AIPs/blob/main/aips/aip-52.md

This AIP proposes to allow sponsored transactions, one in which the gas
payer is not the primary signer of the transaction, to create accounts
for the primary signer if it does not exist. Currently when submitting a
sponsored transaction an account must first exist. This is an
unnecessary friction for using gas fee payer accounts for new accounts.
As it means that the gas fee payer solution must first submit an
independent transaction to the blockchain.

I also cleaned up the tests for fee payer since there were a lot, and they
were excessively complex.

GitOrigin-RevId: 507791c215c58b593af0fb654f9bda500ac05100