[framework] introduce private entry function for auth key rotation (#11969)
This makes it explicit that were rotating an authentication key and
eliminates the need for generating the complex proof the existing
rotation method requires. This is viable because this is a private entry
function that only a transaction can call into that is no scripts or
other functions can call into.

GitOrigin-RevId: fdc041f37e4cb17d2c7f4bb2e0ad784a3f007614