[move-prover] Sharding feature + re-animate Prover.toml (#15775)
Boogie is using exhaustive amounts of memory if run as part of proving aptos-framework (100GB+), which makes CI and local testing fail randomly. This PR introduces a new `--shards` option which splits the Boogie job into multiple shards, if specified. Also re-animates the use of `Prover.toml` in a package directory allowing to set such options.

GitOrigin-RevId: 4125bd8d78786ed26c916c8bf01707f3a3ec9714