[move][closures] End-to-end tests and feature gating  (#15987)
* [move][closures] End-to-end tests and feature gating

Adds move-e2e-test for smoke testing storable closures. Fixes a bug in bytecode verifier and paranoid mode regards computation of closure abilities.

Also introduces a new feature flag ENABLE_FUNCTION_VALUES which allows to turn on and off this feature independent of the bytecode version. This is needed so we can roll out binary version v8 and pass the module upgrade test _before_ this feature is actually enabled. (With other words, guarding with the bytecode version alone doesn't work today.)

* Making the registry example more idiomatic. Function types used for storage always should have a newtype wrapper.

* Fixing feature gating test

* fixing tests

GitOrigin-RevId: 417690b870bedc20c2af350a7ac3269599deb5d2